remove ingredient-suppliers-list , move register and login to auth controller
