+++ routes/authRoutes.js
@@ -2,9 +2,20 @@ const express = require("express");
 const router = express.Router();
 const authControllers = require("../controllers/authControllers");
 const authValidator = require("../utils/validators/authValidator");
+const employeeControllers = require("../controllers/employeeControllers");
+const auth = require("../middlewares/auth");
+const { allowedTo } = require("../controllers/authControllers");
+const customerValidator = require("../utils/validators/customerValidator");
+const customerControllers = require("../controllers/customerControllers");
+
 
 router.post("/login", authValidator.loginValidator, authControllers.login);
-router.post("/register", authValidator.registerValidator, authControllers.register);
-router.post('/employee-account',authValidator.employeeAccountValidator, authControllers.employeeAccount);
+router.post("/register", auth, allowedTo("hr"), authValidator.registerValidator, authControllers.register);
+router.post('/employeeAccount', auth, allowedTo("hr"), employeeControllers.uploadEmployeeImage, employeeControllers.resizeImage, authValidator.employeeAccountValidator, authControllers.employeeAccount);
+
+
+router.post('/CustomerAccount',customerValidator.addCustomerAccount, authControllers.customerAccount);
+router.post('/customerLogin',customerValidator.login, authControllers.customerLogin);
+router.post('/customerRegister', customerControllers.uploadCustomerImage, customerControllers.resizeImage, customerValidator.signup, authControllers.customerSignup);
 
 module.exports = router;
